Moved generic devices to its own file
