+++ bluesky_env/devices/generic.py
@@ -1,0 +1,25 @@
+from ophyd import Component as Cpt
+from ophyd import Device, EpicsMotor, EpicsSignal
+
+
+class WorkPositions(Device):
+    gx = Cpt(EpicsSignal, "{Gov:Robot-Dev:gx}Pos:Work-Pos")
+    gpy = Cpt(EpicsSignal, "{Gov:Robot-Dev:gpy}Pos:Work-Pos")
+    gpz = Cpt(EpicsSignal, "{Gov:Robot-Dev:gpz}Pos:Work-Pos")
+    o = Cpt(EpicsSignal, "{Gov:Robot-Dev:go}Pos:Work-Pos")
+
+
+class MountPositions(Device):
+    gx = Cpt(EpicsSignal, "{Gov:Robot-Dev:gx}Pos:Mount-Pos")
+    py = Cpt(EpicsSignal, "{Gov:Robot-Dev:gpy}Pos:Mount-Pos")
+    pz = Cpt(EpicsSignal, "{Gov:Robot-Dev:gpz}Pos:Mount-Pos")
+    o = Cpt(EpicsSignal, "{Gov:Robot-Dev:go}Pos:Mount-Pos")
+
+
+class GoniometerStack(Device):
+    gx = Cpt(EpicsMotor, "-Ax:GX}Mtr")
+    gy = Cpt(EpicsMotor, "-Ax:GY}Mtr")
+    gz = Cpt(EpicsMotor, "-Ax:GZ}Mtr")
+    o = Cpt(EpicsMotor, "-Ax:O}Mtr")
+    py = Cpt(EpicsMotor, "-Ax:PY}Mtr")
+    pz = Cpt(EpicsMotor, "-Ax:PZ}Mtr")
