+++ start_bs.py
@@ -9,8 +9,8 @@ import os
 from mxtools.governor import _make_governors
 from ophyd.signal import EpicsSignalBase
 EpicsSignalBase.set_defaults(timeout=10, connection_timeout=10)  # new style
-from bluesky_env.devices.auto_center import WorkPositions, TwoClickLowMag, LoopDetector, MountPositions
-from bluesky_env.devices.top_align import TopAlignerFast, TopAlignerSlow, GoniometerStack
+from bluesky_env.devices import (WorkPositions, TwoClickLowMag, LoopDetector, MountPositions, 
+                                 TopAlignerFast, TopAlignerSlow, GoniometerStack)
 
 #12/19 - author unknown. DAMA can help
 """
